pyproject.toml
[project]
name = "euroleague-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.0.1",
    "boto3>=1.38.17",
    "pymongo>=4.13.0",
    "ruff>=0.11.10",
]
